+++ examples/merkle-proofs/src/lib.rs
@@ -18,6 +18,7 @@ sol! {
     error MerkleProofInvalidMultiProofLength();
     error MerkleProofInvalidRootChild();
     error MerkleProofInvalidTotalHashes();
+    error MerkleProofNoLeaves();
 }
 
 #[derive(SolidityError)]
@@ -25,6 +26,7 @@ pub enum VerifierError {
     InvalidProofLength(MerkleProofInvalidMultiProofLength),
     InvalidRootChild(MerkleProofInvalidRootChild),
     InvalidTotalHashes(MerkleProofInvalidTotalHashes),
+    NoLeaves(MerkleProofNoLeaves),
 }
 
 impl core::convert::From<merkle::MultiProofError> for VerifierError {
@@ -43,13 +45,16 @@ impl core::convert::From<merkle::MultiProofError> for VerifierError {
                     MerkleProofInvalidTotalHashes {},
                 )
             }
+            merkle::MultiProofError::NoLeaves => {
+                VerifierError::NoLeaves(MerkleProofNoLeaves {})
+            }
         }
     }
 }
 
 #[entrypoint]
 #[storage]
-struct VerifierContract {}
+struct VerifierContract;
 
 #[public]
 impl VerifierContract {